Create arduino-src-code.c
This a C source code for the question:
"Write C code in Tinkercad for Arduino to control a servo motor using a potentiometer. The servo
should not go beyond 68 degrees and should turn on an LED if it exceeds 68 degrees. The
circuit should have a LCD display displaying how much degree servo motor have rotated. You
need to submit a video of this circuit simulated in Tinkercad and the source code.
"